% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{add_derived_date_info}
\alias{add_derived_date_info}
\title{Add a number of derived time related data}
\usage{
add_derived_date_info(x, light_on, light_off, dst)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{light_on}{Period object with time the light is turned on}

\item{light_off}{Period object with time the light is turned off}

\item{dst}{string either "sync" or "independent". If "sync" light is turned on and off 
when the computer clock are the the timepoints specified in light_on and light_off. If
"independent" light is running on a 24 hour cycle independent of computer time.
These options only matter if the experiment covers a daylight savings time change.}
}
\value{
glucose monitoring data.table (also updated as a side effect)
}
\description{
Add a number of derived time related data
}
\examples{
\dontrun{
to-do
}
}
