% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{add_peak_timers}
\alias{add_peak_timers}
\title{Add various peak timers}
\usage{
add_peak_timers(x, min_peak_duration)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{min_peak_duration}{minimum duration for peak to be counted}
}
\value{
glucose monitoring data.table
}
\description{
Add various peak timers
}
\examples{
\dontrun{
to-do
}
}
