% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cge_analysis.R
\name{analyse_experiment}
\alias{analyse_experiment}
\title{Main wrapper for analyzing a continuous glucose monitoring experiment}
\usage{
analyse_experiment(
  data_file,
  configuration_file,
  out_folder,
  pattern = "Parameters",
  parallel = TRUE,
  reload = TRUE
)
}
\arguments{
\item{data_file}{path to data file}

\item{configuration_file}{path to configuration file, if missing one will be created}

\item{out_folder}{path to folder where results will be saved, if missing one will be created}

\item{pattern}{Regex used to recognize data sheets}

\item{parallel}{logical, should data be loaded in parallel?}

\item{reload}{logical, should previously preprocessed data be reloaded}
}
\value{

}
\description{
Main wrapper for analyzing a continuous glucose monitoring experiment
}
\examples{
\dontrun{
to-do
}
}
