% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{calc_slopes}
\alias{calc_slopes}
\title{Calculate slopes}
\usage{
calc_slopes(x, n_points)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{n_points}{data points used for slope calculation}
}
\value{
glucose monitoring data.table
}
\description{
Calculate slopes
}
\examples{
\dontrun{
to-do
}
}
