% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{config}
\alias{config}
\title{Configuration helper}
\usage{
config(settings, groupings, exclusions)
}
\arguments{
\item{settings}{data.frame with experimental settings}

\item{groupings}{data.frame with sample groupings}

\item{exclusions}{list of data.frames with start and end of excluded timepoints}
}
\value{
configuration object
}
\description{
Configuration helper
}
\examples{
\dontrun{
to-do
}
}
