% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{convert_mgdl_2_molL}
\alias{convert_mgdl_2_molL}
\title{Convert glucose concentrations from mg/dL to mmol/L}
\usage{
convert_mgdl_2_molL(x)
}
\arguments{
\item{x}{x glucose monitoring data.table}
}
\value{
glucose monitoring data.table (also updated as a side effect)
}
\description{
Convert glucose concentrations from mg/dL to mmol/L
}
\examples{
\dontrun{
to-do
}
}
