% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{create_config}
\alias{create_config}
\title{Helper function for creating an analysis configuration file}
\usage{
create_config(path, sample_names, events)
}
\arguments{
\item{path}{Path where config file should be located}

\item{sample_names}{Vector of sample names}

\item{events}{events as reported in the data file}
}
\description{
Helper function for creating an analysis configuration file
}
\examples{
\dontrun{
to-do
}
}
