% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{data_reader}
\alias{data_reader}
\title{data reading helper}
\usage{
data_reader(sheet_name, file, glucose_col, temperature_col, activity_col)
}
\arguments{
\item{sheet_name}{name of sheet to read}

\item{file}{path to data file}

\item{glucose_col}{character, column name for glucose data (typically H)}

\item{temperature_col}{character, column name for temperature data}

\item{activity_col}{character, column name for activity data}
}
\value{
data.table
}
\description{
data reading helper
}
\examples{
\dontrun{
to-do
}

}
