% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{exclude_missing}
\alias{exclude_missing}
\title{Mark timepoints without glucose data as excluded}
\usage{
exclude_missing(x)
}
\arguments{
\item{x}{glucose monitoring data.table}
}
\value{
glucose monitoring data.table
}
\description{
Mark timepoints without glucose data as excluded
}
\examples{
\dontrun{
to-do
}
}
