% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{exclude_timepoints}
\alias{exclude_timepoints}
\title{Exclude timepoints}
\usage{
exclude_timepoints(x, exclusions, invert)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{exclusions}{data.frame-like object with POSIXct columns Start and End denoting exclusion intervals}
}
\value{
x glucose monitoring data.table (also updated as a side effect)
}
\description{
Exclude timepoints
}
\examples{
\dontrun{
to-do
}
}
