% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{filter_max_missing}
\alias{filter_max_missing}
\title{Filter missing data}
\usage{
filter_max_missing(x, idx, min_frac_summaries)
}
\arguments{
\item{x}{a vector}

\item{idx}{logical}

\item{min_frac_summaries}{minimum average of idx}
}
\value{
x or NA
}
\description{
Filter missing data
}
\examples{
\dontrun{
to-do
}
}
