% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{find_baseline}
\alias{find_baseline}
\title{Find baseline glucose expression}
\usage{
find_baseline(x, baseline_window, max_missing_baseline)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{baseline_window}{number of timepoints to include in baseline calculations}

\item{max_missing_baseline}{maximum number of excluded datapoints before baseline is set to NA}
}
\value{
glucose monitoring data.table
}
\description{
Find baseline glucose expression
}
\examples{
\dontrun{
to-do
}
}
