% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{find_baseline_ignore_high}
\alias{find_baseline_ignore_high}
\title{Find baseline glucose expression, ignoring high glucose values}
\usage{
find_baseline_ignore_high(x, baseline_window, max_missing_baseline)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{baseline_window}{number of timepoints to include in baseline calculations}

\item{max_missing_baseline}{maximum number of excluded datapoints before baseline is set to NA}
}
\value{
glucose monitoring data.table
}
\description{
Find baseline glucose expression, ignoring high glucose values
}
\examples{
\dontrun{
to-do
}
}
