% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{find_peaks_and_nadirs}
\alias{find_peaks_and_nadirs}
\title{Find peaks and nadirs}
\usage{
find_peaks_and_nadirs(x, max_min_window)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{max_min_window}{integer, window to scan for local peaks/nadirs. Must be odd.}
}
\value{
glucose monitoring data.table
}
\description{
Find peaks and nadirs
}
\examples{
\dontrun{
to-do
}
}
