% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{fix_date}
\alias{fix_date}
\title{Update Date column to include time zone information}
\usage{
fix_date(x, tz)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{tz}{timezone to use}
}
\value{
glucose monitoring data.table (also updated as a side effect)
}
\description{
Update Date column to include time zone information
}
\examples{
\dontrun{
to-do
}
}
