% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{get_exclusions}
\alias{get_exclusions}
\title{Get exclusions from CGM experiment}
\usage{
get_exclusions(x, sample)
}
\arguments{
\item{x}{cgm_experiment}

\item{sample}{sample name}
}
\value{
tibble with exclusion times
}
\description{
Get exclusions from CGM experiment
}
\examples{
\dontrun{
to-do
}
}
