% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{get_option}
\alias{get_option}
\title{Get options from CGM experiment}
\usage{
get_option(x, option)
}
\arguments{
\item{x}{cgm_experiment}

\item{option}{option to extract}
}
\value{
value of the selected option
}
\description{
Get options from CGM experiment
}
\examples{
\dontrun{
to-do
}
}
