% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{get_profiles}
\alias{get_profiles}
\title{Get individual and group-wise profiles}
\usage{
get_profiles(cge, stat, low, high, step, as_percent, subset_expression = TRUE)
}
\arguments{
\item{cge}{cgm_experiment object}

\item{stat}{unevaluated expression that will be performed within x}

\item{low}{lowest value of profile range}

\item{high}{highest value of profile range}

\item{step}{numeric, interval between bins}

\item{as_percent}{logical, report results as percent instead of minutes per hour}

\item{subset_expression}{expression used to subset the data (often (peak | nadir))}
}
\value{
data.frame
}
\description{
Get individual and group-wise profiles
}
\examples{
\dontrun{
to-do
}
}
