% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{get_sample_kinetics}
\alias{get_sample_kinetics}
\title{get kinetics calculations of a single sample}
\usage{
get_sample_kinetics(x, excursion_high, min_peak_duration, datapoints_for_slope)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{excursion_high}{increase above baseline before excursion is tagged}

\item{min_peak_duration}{mimumim peak duration}

\item{datapoints_for_slope}{datapoints used for slope calculation}
}
\value{

}
\description{
get kinetics calculations of a single sample
}
\examples{
\dontrun{
to-do
}
}
