% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{get_spring_constants}
\alias{get_spring_constants}
\title{Get spring constants}
\usage{
get_spring_constants(kinetics_all)
}
\arguments{
\item{kinetics_all}{data.table with kinetics data}
}
\value{
list with spring constants
}
\description{
Get spring constants
}
\examples{
\dontrun{
to-do
}
}
