% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{glucose_data}
\alias{glucose_data}
\title{Glucose data helper}
\usage{
glucose_data(x)
}
\arguments{
\item{x}{list of data.tables with (potentially a subset of) the columns 
needed for the full analysis}
}
\value{
a cgm_glucose_data object
}
\description{
Glucose data helper
}
\examples{
\dontrun{
to-do
}
}
