% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{glucose_statistics}
\alias{glucose_statistics}
\title{Get glucose statistics}
\usage{
glucose_statistics(cge)
}
\arguments{
\item{cge}{cgm_experiment object}
}
\value{
data.table with summaries
}
\description{
Get glucose statistics
}
\examples{
\dontrun{
to-do
}
}
