% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{glucose_statistics_sample}
\alias{glucose_statistics_sample}
\title{Get sample glucose statistics}
\usage{
glucose_statistics_sample(by, sample_data, min_frac_summaries)
}
\arguments{
\item{by}{character, columns to aggregate by}

\item{sample_data}{glucose monitoring data.table}

\item{min_frac_summaries}{minimum fraction of included data for calculating summaries}
}
\value{
data.table with summaries
}
\description{
Get sample glucose statistics
}
\examples{
\dontrun{
to-do
}
}
