% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{isoglycemic_statistics}
\alias{isoglycemic_statistics}
\title{Get isoglycemic statistics}
\usage{
isoglycemic_statistics(cge)
}
\arguments{
\item{cge}{cgm_experiment object}
}
\value{
data.table with summaries
}
\description{
Get isoglycemic statistics
}
\examples{
\dontrun{
to-do
}
}
