% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{kinetics_counts_single}
\alias{kinetics_counts_single}
\title{Get counts of kinetics observations}
\usage{
kinetics_counts_single(by, kinetics)
}
\arguments{
\item{by}{character, columns to aggregate by}

\item{kinetics}{table with kinetics data}
}
\value{
data.table with number kinetics observations
}
\description{
Get counts of kinetics observations
}
