% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{kinetics_statistics}
\alias{kinetics_statistics}
\title{Get kinetics statistics}
\usage{
kinetics_statistics(cge)
}
\arguments{
\item{cge}{cgm_experiment object}
}
\value{
data.table with summaries
}
\description{
Get kinetics statistics
}
\examples{
\dontrun{
to-do
}
}
