% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{kinetics_statistics_single}
\alias{kinetics_statistics_single}
\title{Get statistics on kinetics parameters}
\usage{
kinetics_statistics_single(by, column, kinetics)
}
\arguments{
\item{by}{character, columns to aggregate by}

\item{column}{column to calculate statistics on}

\item{kinetics}{table with kinetics data}
}
\value{
data.table with kinetics summaries
}
\description{
Get statistics on kinetics parameters
}
