% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{linear_imputation}
\alias{linear_imputation}
\title{Linear interpolation of glucose for small gaps}
\usage{
linear_imputation(x, column, max_gap)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{column}{character, column to impute}

\item{max_gap}{maximum number of missing datapoints to impute}
}
\value{
x glucose monitoring data.table (NOT updated as a side effect)
}
\description{
Linear interpolation of glucose for small gaps
}
\examples{
\dontrun{
to-do
}
}
