% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{make_breaks}
\alias{make_breaks}
\title{Make breaks}
\usage{
make_breaks(x, low, high, step, divide_by)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{low}{mimimum value}

\item{high}{maximum value}

\item{step}{interval between bins}

\item{divide_by}{number to divide by. If data is sampled every minute and 
this is set to 60, then results are returned as per hour.}
}
\value{

}
\description{
Make breaks
}
