% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{make_profile}
\alias{make_profile}
\title{Calculate profiles}
\usage{
make_profile(
  x,
  by_row,
  by_col,
  stat,
  low,
  high,
  min_frac_summaries,
  subset_expression,
  update_names = TRUE
)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{by_row}{what should be in the rows (often Light_on)}

\item{by_col}{what should be in the rows (often Day or Week)}

\item{stat}{unevaluated expression that will be performed within x}

\item{low}{lowest value of profile range}

\item{high}{highest value of profile range}

\item{min_frac_summaries}{minimum fraction of data needed before profiling is performed}

\item{subset_expression}{expression used to subset the data (often (peak | nadir))}

\item{update_names}{logical, should column names have the name of the variable prepended?}
}
\value{
data.table
}
\description{
Calculate profiles
}
\examples{
\dontrun{
to-do
}
}
