% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{multi_peak_kinetics}
\alias{multi_peak_kinetics}
\title{Single peak kinetics}
\usage{
multi_peak_kinetics(x, excursion_high)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{excursion_high}{increase above baseline before excursion is tagged}
}
\value{
kinetics data.table
}
\description{
Single peak kinetics
}
\examples{
\dontrun{
to-do
}
}
