% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{names<-.cgm_experiment}
\alias{names<-.cgm_experiment}
\title{The Names of an Object}
\usage{
\method{names}{cgm_experiment}(x) <- value
}
\arguments{
\item{x}{cgm_experiment object}

\item{new_names}{vector of new names}
}
\value{

}
\description{
The Names of an Object
}
\examples{
\dontrun{
to-do
}
}
