% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{names.cgm_experiment}
\alias{names.cgm_experiment}
\title{The Names of an Object}
\usage{
\method{names}{cgm_experiment}(x)
}
\arguments{
\item{x}{cgm_experiment object}
}
\value{
characters
}
\description{
The Names of an Object
}
\examples{
\dontrun{
to-do
}
}
