% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{new_cgm_experiment}
\alias{new_cgm_experiment}
\title{Continous glucose monitoring experiment constructor}
\usage{
new_cgm_experiment(glucose_data, config)
}
\arguments{
\item{glucose_data}{gcm_data object}

\item{config}{gcm_config object}
}
\value{
cgm_experiment object
}
\description{
Continous glucose monitoring experiment constructor
}
\examples{
\dontrun{
to-do
}
}
