% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{new_config}
\alias{new_config}
\title{Configuration constructor}
\usage{
new_config(settings, groupings, exclusions)
}
\arguments{
\item{settings}{data.frame with experimental settings}

\item{groupings}{data.frame with sample groupings}

\item{exclusions}{list of data.frames with start and end of excluded timepoints}
}
\value{
cgm_config object
}
\description{
Configuration constructor
}
\examples{
\dontrun{
to-do
}
}
