% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{new_glucose_data}
\alias{new_glucose_data}
\title{Glucose data constructor}
\usage{
new_glucose_data(x = data.table::data.table())
}
\arguments{
\item{x}{list of data.tables with (potentially a subset of) the columns 
needed for the full analysis}
}
\value{
a cgm_glucose_data object
}
\description{
Glucose data constructor
}
\examples{
\dontrun{
to-do
}
}
