% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{other_statistics_aggregate}
\alias{other_statistics_aggregate}
\title{Get average temperature/activity statistics}
\usage{
other_statistics_aggregate(by, sample_statistics)
}
\arguments{
\item{by}{character, columns to aggregate by}

\item{sample_statistics}{list of results generated by glucose_statistics_sample}
}
\value{
data.table with summaries
}
\description{
Get average temperature/activity statistics
}
\examples{
\dontrun{
to-do
}
}
