% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{other_statistics_sample}
\alias{other_statistics_sample}
\title{Get sample temperature/activity statistics}
\usage{
other_statistics_sample(by, sample_data, min_frac_summaries, column)
}
\arguments{
\item{by}{character, columns to aggregate by}

\item{sample_data}{glucose monitoring data.table}

\item{min_frac_summaries}{minimum fraction of included data for calculating summaries}

\item{column}{column to be summarized}
}
\value{
data.table with summaries
}
\description{
Get sample temperature/activity statistics
}
\examples{
\dontrun{
to-do
}
}
