% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{peak_statistics}
\alias{peak_statistics}
\title{Get peak statistics}
\usage{
peak_statistics(cge)
}
\arguments{
\item{cge}{cgm_experiment object}
}
\value{
data.table with summaries
}
\description{
Get peak statistics
}
\examples{
\dontrun{
to-do
}
}
