% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{peak_statistics_aggregate}
\alias{peak_statistics_aggregate}
\title{Get average peak statistics}
\usage{
peak_statistics_aggregate(by, sample_statistics)
}
\arguments{
\item{by}{character, columns to aggregate by}

\item{sample_statistics}{list of results generated by peak_statistics_sample}
}
\value{
data.table with summaries
}
\description{
Get average peak statistics
}
\examples{
\dontrun{
to-do
}
}
