% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{peak_statistics_sample}
\alias{peak_statistics_sample}
\title{Get sample peak statistics}
\usage{
peak_statistics_sample(by, sample_data, min_frac_summaries)
}
\arguments{
\item{by}{character, columns to aggregate by}

\item{sample_data}{glucose monitoring data.table}

\item{min_frac_summaries}{minimum fraction of included data for calculating summaries}
}
\value{
data.table with summaries
}
\description{
Get sample peak statistics
}
\examples{
\dontrun{
to-do
}
}
