% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{preprocess_samples}
\alias{preprocess_samples}
\title{Preprocess all samples in an experiment}
\usage{
preprocess_samples(cge)
}
\arguments{
\item{cge}{cgm_experiment object}
}
\value{
cgm_experiment object
}
\description{
Preprocess all samples in an experiment
}
\examples{
\dontrun{
to-do
}
}
