% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{print.cgm_experiment}
\alias{print.cgm_experiment}
\title{Print Values}
\usage{
\method{print}{cgm_experiment}(x, ...)
}
\arguments{
\item{x}{cgm_experiment object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print Values
}
\examples{
\dontrun{
to-do
}
}
