% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read_config}
\alias{read_config}
\title{Read configurations from config file}
\usage{
read_config(file, samples)
}
\arguments{
\item{file}{path to configurations file}

\item{samples}{names of samples}
}
\value{
cgm_config object
}
\description{
Read configurations from config file
}
\examples{
\dontrun{
to-do
}
}
