% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read_data}
\alias{read_data}
\title{Read continuous glucose monitoring data}
\usage{
read_data(file, pattern = "Parameters", parallel = FALSE)
}
\arguments{
\item{file}{file with continuous glucose monitoring data}

\item{pattern}{regex to recognize sheets containing data}

\item{parallel}{logical, should data be loaded in parallel?}
}
\value{
cgm_glucose_data object
}
\description{
Read continuous glucose monitoring data
}
\examples{
\dontrun{
to-do
}
}
