% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read_exclusion}
\alias{read_exclusion}
\title{Read exclusions}
\usage{
read_exclusion(sheet, file)
}
\arguments{
\item{sheet}{name of sheet to read}

\item{file}{path to config file}
}
\value{
list of exclusions
}
\description{
Read exclusions
}
\examples{
\dontrun{
to-do
}
}
