% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read_groupings}
\alias{read_groupings}
\title{Read sample groupings}
\usage{
read_groupings(file)
}
\arguments{
\item{file}{path to config file}
}
\value{
list of groupings
}
\description{
Read sample groupings
}
\examples{
\dontrun{
to-do
}
}
