% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read_settings}
\alias{read_settings}
\title{Read settings}
\usage{
read_settings(file)
}
\arguments{
\item{file}{path to config file}
}
\value{
list of settings
}
\description{
Read settings
}
\examples{
\dontrun{
to-do
}
}
