% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{run_standard_preprocess_pipeline}
\alias{run_standard_preprocess_pipeline}
\title{Run the standard pipeline for a single sample}
\usage{
run_standard_preprocess_pipeline(sample_id, cge)
}
\arguments{
\item{sample_id}{sample to be processed}

\item{cge}{cgm data object}
}
\value{
glucose monitoring data.table
}
\description{
Run the standard pipeline for a single sample
}
\examples{
\dontrun{
to-do
}
}
