% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{secondary_filtering}
\alias{secondary_filtering}
\title{Secondary filtering of peaks}
\usage{
secondary_filtering(x, peak_ratio)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{peak_ratio}{minimum ratio between peak value and minimum obtained since last peak}
}
\value{
glucose monitoring data.table
}
\description{
Secondary filtering of peaks
}
\examples{
\dontrun{
to-do
}
}
