% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{select_excursions}
\alias{select_excursions}
\title{Select excursions to include in kinetics calculations}
\usage{
select_excursions(x, min_peak_duration)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{min_peak_duration}{integer, minimum excursion duration before peak is used for kinetics calculations}
}
\value{
glucose monitoring data.table
}
\description{
Select excursions to include in kinetics calculations
}
\examples{
\dontrun{
to-do
}
}
