% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{smooth_background}
\alias{smooth_background}
\title{Smooth background}
\usage{
smooth_background(x, baseline_window)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{baseline_window}{number of timepoints to include in baseline calculations}
}
\value{
glucose monitoring data.table
}
\description{
Smooth background
}
\examples{
\dontrun{
to-do
}
}
