% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{[.cgm_experiment}
\alias{[.cgm_experiment}
\title{Extract or Replace Parts of an Object}
\usage{
\method{[}{cgm_experiment}(x, i)
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s).}

\item{i}{indices specifying elements to extract}
}
\value{
object of same type as x
}
\description{
Extract or Replace Parts of an Object
}
\examples{
\dontrun{
to-do
}
}
