% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{tag_events}
\alias{tag_events}
\title{Tag events}
\usage{
tag_events(x, event, before, after)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{event}{character (regex) to denoting event(s) to detect}

\item{before}{int, minutes before event to tag}

\item{after}{int, minutes after event to tag}
}
\value{
glucose monitoring data.table
}
\description{
Tag events
}
\examples{
\dontrun{
to-do
}
}
