% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{tag_excursions}
\alias{tag_excursions}
\title{Tag excursions}
\usage{
tag_excursions(x, excursion_low, excursion_high)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{excursion_low}{increase below baseline before excursion is tagged}

\item{excursion_high}{increase above baseline before excursion is tagged}
}
\value{
glucose monitoring data.table
}
\description{
Tag excursions
}
\examples{
\dontrun{
to-do
}
}
