% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{tag_multi_nadirs}
\alias{tag_multi_nadirs}
\title{Tag excursion with multiple nadirs}
\usage{
tag_multi_nadirs(x)
}
\arguments{
\item{x}{glucose monitoring data.table}
}
\value{
glucose monitoring data.table
}
\description{
Tag excursion with multiple nadirs
}
\examples{
\dontrun{
to-do
}
}
