% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{tag_multi_peaks}
\alias{tag_multi_peaks}
\title{Tag excursion with multiple peaks}
\usage{
tag_multi_peaks(x)
}
\arguments{
\item{x}{glucose monitoring data.table}
}
\value{
glucose monitoring data.table
}
\description{
Tag excursion with multiple peaks
}
\examples{
\dontrun{
to-do
}
}
