% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_glucose_data.R
\name{uniformize_sample_rate}
\alias{uniformize_sample_rate}
\title{Update data to be sampled every minute}
\usage{
uniformize_sample_rate(x, sample_rate = 60)
}
\arguments{
\item{x}{glucose monitoring data.table}

\item{sample_rate}{sampling rate in seconds}
}
\value{
glucose monitoring data.table (NOT updated as a side effect)
}
\description{
Update data to be sampled every minute
}
\examples{
\dontrun{
to-do
}
}
