% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{update_names}
\alias{update_names}
\title{The Names of an Object}
\usage{
update_names(x, value)
}
\arguments{
\item{x}{cgm_experiment object}

\item{value}{vector of new names}
}
\value{

}
\description{
The Names of an Object
}
\examples{
\dontrun{
to-do
}
}
