% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{validate_cgm_experiment}
\alias{validate_cgm_experiment}
\title{Continous glucose monitoring experiment validator}
\usage{
validate_cgm_experiment(x)
}
\arguments{
\item{x}{cgm_experiment object}
}
\value{
cgm_experiment object
}
\description{
Continous glucose monitoring experiment validator
}
\examples{
\dontrun{
to-do
}
}
