% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{validate_config}
\alias{validate_config}
\title{Configuration validator}
\usage{
validate_config(x)
}
\arguments{
\item{x}{configuration object}
}
\value{
cgm_config object
}
\description{
Configuration validator
}
\examples{
\dontrun{
to-do
}
}
