% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{validate_exclusions}
\alias{validate_exclusions}
\title{Exclusions validator}
\usage{
validate_exclusions(x)
}
\arguments{
\item{x}{exclusions part of cgm_config object}
}
\value{
exclusions part of cgm_config object
}
\description{
Exclusions validator
}
\examples{
\dontrun{
to-do
}
}
