% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{validate_glucose_data}
\alias{validate_glucose_data}
\title{Glucose data validator}
\usage{
validate_glucose_data(x)
}
\arguments{
\item{x}{a cgm_glucose_data object}
}
\value{
a cgm_glucose_data object
}
\description{
Glucose data validator
}
\examples{
\dontrun{
to-do
}
}
