% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{validate_grouping}
\alias{validate_grouping}
\title{Groupings validator}
\usage{
validate_grouping(x)
}
\arguments{
\item{x}{groupings part of cgm_config object}
}
\value{
groupings part of cgm_config object
}
\description{
Groupings validator
}
\examples{
\dontrun{
to-do
}
}
