% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_data_classes.R
\name{validate_settings}
\alias{validate_settings}
\title{Settings validator}
\usage{
validate_settings(x)
}
\arguments{
\item{x}{settings part of a cgm_config object}
}
\value{
settings part of a cgm_config object
}
\description{
Settings validator
}
\examples{
\dontrun{
to-do
}
}
