% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{data_reader}
\alias{data_reader}
\title{data reading helper}
\usage{
data_reader(sheet_name, file)
}
\arguments{
\item{sheet_name}{name of sheet to read}

\item{file}{path to data file}
}
\value{
data.table
}
\description{
data reading helper
}
\examples{
\dontrun{
to-do
}
}
