% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{prepare_experiment}
\alias{prepare_experiment}
\title{Helper script for preparing for data analysis}
\usage{
prepare_experiment(
  data_file,
  configuration_file,
  pattern = "Parameters",
  parallel = TRUE
)
}
\arguments{
\item{data_file}{path to data file}

\item{configuration_file}{path to configuration file, if missing one will be created}

\item{pattern}{Regex used to recognize data sheets}

\item{parallel}{logical, should data be loaded in parallel?}
}
\value{
cgm_experiment object
}
\description{
Helper script to prepare data generated using this data generation platform.
}
\examples{
\dontrun{
to-do
}
}
